Design(#30-battery-view): dashedLine을 위한 Line Shape 추가
