+++ SeeSaw/Extensions/Shape/Line.swift
@@ -1,0 +1,17 @@
+//
+//  Line.swift
+//  SeeSaw
+//
+//  Created by 이안진 on 2023/05/19.
+//
+
+import SwiftUI
+
+struct Line: Shape {
+    func path(in rect: CGRect) -> Path {
+        var path = Path()
+        path.move(to: CGPoint(x: 0, y: 0))
+        path.addLine(to: CGPoint(x: rect.width, y: 0))
+        return path
+    }
+}
